+++ cypress/e2e/33_orgBountyStatusFilter.cy.ts
@@ -251,7 +251,7 @@ describe('filter by status for org bounty', () => {
 
     cy.contains('Status').first().click();
     cy.wait(1000);
-    
+
     cy.logout(org.loggedInAs);
   });
 });
